Move to the native approach
