+++ next-frontend/src/sessionOrSignIn.ts
@@ -1,12 +1,9 @@
-import { auth, signOut } from "@/auth";
+import { auth, signIn } from "@/auth";
 
 export async function sessionOrSignIn() {
   const session = await auth();
   if (session?.error === "RefreshTokenError") {
-    // Refresh failed (revoked / expired / race outside the Doorkeeper grace
-    // window). Drop the session so the user is forced through a fresh login
-    // rather than continuing with a stale access_token.
-    await signOut();
+    await signIn(); // Force sign in to get a new set of access and refresh tokens
   }
   return session;
 }
